pyproject.toml
[project]
name = "donatio-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.4",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.3",
    "django-allauth>=65.9.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "requests>=2.32.4",
    "ruff>=0.12.1",
]
